requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distro==1.8.0
Django==4.2.7
exceptiongroup==1.1.3
fonttools==4.44.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
image==1.5.33
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.1
multidict==6.0.4
numpy==1.26.1
openai==1.2.3
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
plotly==5.18.0
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
snakeviz==2.2.0
sniffio==1.3.0
sqlparse==0.4.4
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
